+++ maps/src/com/mobidevelop/maps/MapObject.java
@@ -26,21 +26,21 @@ public interface MapObject {
 
 	public abstract void setName(String name);
 	
-	public abstract int getX();
+	public abstract float getX();
 
-	public abstract void setX(int x);
+	public abstract void setX(float x);
 
-	public abstract int getY();
+	public abstract float getY();
 
-	public abstract void setY(int y);
+	public abstract void setY(float y);
 
-	public abstract int getWidth();
+	public abstract float getWidth();
 
-	public abstract void setWidth(int width);
+	public abstract void setWidth(float width);
 
-	public abstract int getHeight();
+	public abstract float getHeight();
 
-	public abstract void setHeight(int height);
+	public abstract void setHeight(float height);
 
 	public abstract float getRotation();
